Expanded form and data
Added more inputs to the form, I currently have everything as dropdowns,
may change to radio buttons, we'll see. The location seems to be messed
up when pulling random pets, so I think I'm going to make a list of
shelters for us to pull and then use a random number generator to decide
which one to pull an animal from. anyway, etc, etc.